+++ src/components/common/Hero.jsx
