+++ src/components/events/EventHeroSection.jsx
