+++ src/components/home/EventsSection.jsx
